# Greatly improve the performance of form_tree when regridding.