@JiakunYan 2023-05-18
More details:

I also implemented tracing for HPX. The blue bar shows the number of messages sent every 0.1 seconds. The orange line shows the total bytes sent every 0.1 seconds.

Rank 0: 
Before the change: 
<img src="https://github.com/STEllAR-GROUP/octotiger/assets/29711311/28f3bb66-9ca0-4655-b49c-1c118ef4535e" width="400" height="300">

After the change:
<img src="https://github.com/STEllAR-GROUP/octotiger/assets/29711311/08d2b33e-fab8-49f6-ad56-620e5fed978e" width="400" height="300">

Rank 21:
Before the change: 
<img src="https://github.com/STEllAR-GROUP/octotiger/assets/29711311/265fabb3-0968-46bc-862e-2248b5028ad4" width="400" height="300">

After the change:
<img src="https://github.com/STEllAR-GROUP/octotiger/assets/29711311/e4895685-ee81-4a81-a55f-e49ceb00e06f" width="400" height="300">
     
Rank 31:
Before the change: 
<img src="https://github.com/STEllAR-GROUP/octotiger/assets/29711311/c735b925-507b-4911-9524-0f43130d3d02" width="400" height="300">

After the change:
<img src="https://github.com/STEllAR-GROUP/octotiger/assets/29711311/2a107dcb-0ccf-40ea-9c66-331fea092727" width="400" height="300">


There are time durations when Octo-tiger is sending almost no messages. With some prints, I found Octo-Tiger was performing “form tree” at that time. There are also small spikes of messages in these “form tree” duration. The time of these spikes changes from rank to rank. Therefore, I think Octo-Tiger is doing poorly in parallelizing the “form tree” task between ranks, and this PR greatly improves it.