@JiakunYan 2023-05-22
@hkaiser Actually I am curious: what does it mean to have a `sync` launch policy for dataflow? Based on my understanding, the dataflow just creates a thread that will be ready to run once all the input futures are ready?